ASKAPSDP-1477 - Beam log output for simager
This change allows the user to write out a log of the restoring beams 
for the individual channels. This makes use of the BeamLogger in 
imageacess - I have removed the writing out of individual images in 
the BeamLogger output, since we don't do it that way any more.

This change also means the restored and psf.image cubes are written 
only when the restore flag is set (as that is the only time they are 
created).