ASKAPSDP-2589 - Implementation of FITS output of noise etc images
This merges to trunk the work done on features/selavy-dev to enable 
direct FITS output of the noise/mean/snr/threshold images created by 
Selavy. 
This improves the FITS image accessor interface, allowing the write 
function to write a subset of the image in all dimensions. 
The interfaces to DistributedImageWriter & ImageWriter in Selavy have 
changed to allow the passing of parsets (this instantiates the correct 
image accessor).