ASKAPSDP-2882 Merge of the development branch of the CDS service. (details in features/cds-integration-redux) This implements checkin implements the accessor to the calibration data service. Currently the Service source is a stub in Base/accessors, that is over-ridden by an implementation in Components/Services/Calibration. The accessor also lives there - but to avoid replication I have used a client that was developed by Ben that lives in cpdataservices. I could completely replicate this or move it - but this is also where some other classes live and some tools (like ingest) depend upon it.
But I did move some of the definitions from cpdataservices to Base/accessors as these were complete replications and it seemed dangerous. Also I <think> there was an error in IceMapper that was causing a SEGV when I attempted to pull a Bandpass Solution via Ice. - I think that is now fixed - but some implementation tickets that will follow this will see whether that is actually the case

Steve