Place generated files under binary directory
Generated files should go to the binary directory rather than cluttering
the source directory. In the latter case we end up with "dirty"
repositories after a compilation, which is not a nice thing to have.

I also took the chance of de-generating askap_accessors.h, as it didn't
really need any code generation at all -- it's already self-contained.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>